Remove filename support
I'm planning to also download the input and the amount of customisation
required to make the script support specifying the name of the input
file would be too cumbersome for too little benefit. Currently, the
script is just for personal use and I know I'd not use this feature. In
the future it may be added. I don't oppose, but I ought to see demand
(someone asking for it or even contributing themselves).